requirements.txt
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
torch-geometric>=2.5.2
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.22.0
tensorboard>=2.16.0
pyyaml>=6.0.1
tqdm>=4.66.0
streamlit>=1.37.0
gradio>=4.44.0
umap-learn>=0.5.6
scipy>=1.12.0
requests>=2.32.0

